+++ src/app/(normal)/channel/[id]/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
 
-import { ProfileDetailPage } from '@/app/(normal)/profile/pages/DetailPage.js';
+import { ChannelDetailPage } from '@/app/(normal)/channel/pages/DetailPage.js';
 import { KeyType, type SourceInURL } from '@/constants/enum.js';
 import { createSiteMetadata } from '@/helpers/createSiteMetadata.js';
 import { isBotRequest } from '@/helpers/isBotRequest.js';
@@ -27,5 +27,5 @@ export async function generateMetadata({ params, searchParams }: Props): Promise
 export default function Page(props: Props) {
     if (isBotRequest()) return null;
 
-    return <ProfileDetailPage {...props} />;
+    return <ChannelDetailPage {...props} />;
 }
