+++ src/services/getChannelById.ts
@@ -8,7 +8,7 @@ export function getChannelById(source: SocialPlatform, channelId: string) {
         case SocialPlatform.Lens:
             return null;
         case SocialPlatform.Farcaster:
-            // in fact, channelId is channelHandle, we use it handle in the url
+            // in fact, channelId is channelHandle
             return FarcasterSocialMediaProvider.getChannelByHandle(channelId);
         case SocialPlatform.Twitter:
             return null;
